@@ -28,66 +28,67 @@
     "serve:aot": "cd tools && npm run cli:serve:aot"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.9",
-    "@angular-devkit/build-ng-packagr": "^0.803.9",
-    "@angular/animations": "^8.2.10",
-    "@angular/cdk": "^8.2.3",
-    "@angular/cli": "^8.3.26",
-    "@angular/common": "^8.2.10",
-    "@angular/compiler": "^8.2.10",
-    "@angular/compiler-cli": "^8.2.10",
-    "@angular/core": "^8.2.10",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.10",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.10",
-    "@angular/platform-browser-dynamic": "^8.2.10",
-    "@angular/router": "^8.2.10",
-    "@compodoc/compodoc": "^1.1.10",
-    "@firestitch/api": "^8.0.1",
-    "@firestitch/autocomplete": "^8.2.3",
-    "@firestitch/autocomplete-chips": "^8.0.0",
-    "@firestitch/checkboxgroup": "^7.1.10",
-    "@firestitch/clear": "^7.1.0",
+    "@angular-devkit/build-angular": "~0.901.1",
+    "@angular-devkit/build-ng-packagr": "~0.901.1",
+    "@angular/animations": "^9.1.1",
+    "@angular/cdk": "^9.2.0",
+    "@angular/cli": "^9.1.1",
+    "@angular/common": "^9.1.1",
+    "@angular/compiler": "^9.1.1",
+    "@angular/compiler-cli": "^9.1.1",
+    "@angular/core": "^9.1.1",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "^9.1.1",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^9.2.0",
+    "@angular/platform-browser": "^9.1.1",
+    "@angular/platform-browser-dynamic": "^9.1.1",
+    "@angular/router": "^9.1.1",
+    "@compodoc/compodoc": "^1.1.11",
+    "@firestitch/api": "^8.1.8",
+    "@firestitch/autocomplete": "^8.7.0",
+    "@firestitch/autocomplete-chips": "^8.3.4",
+    "@firestitch/checkboxgroup": "^8.1.1",
+    "@firestitch/clear": "^8.1.0",
     "@firestitch/common": "^8.0.1",
-    "@firestitch/datepicker": "^8.0.1",
-    "@firestitch/drawer": "^8.3.0",
-    "@firestitch/example": "^8.0.0",
+    "@firestitch/datepicker": "^8.2.2",
+    "@firestitch/dialog": "^8.1.1",
+    "@firestitch/drawer": "^8.3.3",
+    "@firestitch/example": "^9.0.0",
     "@firestitch/iframe": "^8.0.0",
-    "@firestitch/label": "^7.1.9",
+    "@firestitch/label": "^8.2.1",
     "@firestitch/menu": "^9.0.2",
     "@firestitch/message": "^8.1.0",
-    "@firestitch/dialog": "^8.1.1",
-    "@firestitch/password": "^7.0.4",
-    "@firestitch/phone": "^7.2.1",
-    "@firestitch/prompt": "^8.3.0",
-    "@firestitch/radiogroup": "^7.2.5",
-    "@firestitch/scroll-picker": "^7.0.3",
-    "@firestitch/skeleton": "^7.1.2",
-    "codelyzer": "^5.1.2",
+    "@firestitch/password": "^8.1.0",
+    "@firestitch/phone": "^8.1.0",
+    "@firestitch/prompt": "^8.4.2",
+    "@firestitch/radiogroup": "^8.1.1",
+    "@firestitch/scroll-picker": "^8.0.0",
+    "@firestitch/skeleton": "^8.1.4",
+    "codelyzer": "^5.2.2",
     "cpx": "^1.5.0",
-    "date-fns": "^2.4.1",
+    "date-fns": "^2.12.0",
     "hammerjs": "^2.0.8",
-    "highlight.js": "^9.15.10",
-    "imask": "^5.2.1",
+    "highlight.js": "^9.18.1",
+    "imask": "^6.0.3",
     "iscroll": "^5.2.0",
     "lodash-es": "^4.17.15",
-    "ng-packagr": "^5.6.1",
-    "ngx-toastr": "^11.1.0",
+    "ng-packagr": "^9.0.0",
+    "ngx-highlightjs": "^4.1.0-beta",
+    "ngx-toastr": "^12.0.1",
     "npm-run-all": "^4.1.5",
-    "rxjs": "^6.5.3",
-    "tsickle": "^0.37.0",
-    "tslint": "^5.20.0",
-    "typescript": "~3.5.3",
-    "zone.js": "^0.10.2"
+    "rxjs": "^6.5.5",
+    "tslint": "^6.1.1",
+    "typescript": "^3.8.3",
+    "zone.js": "~0.10.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "peerDependencies": {
-    "rxjs": ">=6.3.0",
-    "@firestitch/common": ">=7.0.2",
-    "@firestitch/drawer": ">=8.0.0"
+    "@firestitch/common": ">=7.0.2"
+  },
+  "dependencies": {
+    "tslib": "^1.10.0"
   }
 }
